refactor: handleConfirmDelete 함수 분리 (#142)
- handleConfirmDelete 함수를 useScheduleDelete 훅으로 분리
